DCM_data_pretreatment/requirement.txt
pydicom
numpy
pypng
cv2
time